+++ src/TortoiseProc/refloglist.cpp
@@ -78,7 +78,7 @@ void CRefLogList::OnLvnGetdispinfoLoglist(NMHDR *pNMHDR, LRESULT *pResult)
 	// By default, clear text buffer.
 	lstrcpyn(pItem->pszText, L"", pItem->cchTextMax);
 
-	bool bOutOfRange = pItem->iItem >= m_arShownList.size();
+	bool bOutOfRange = pItem->iItem >= (int)m_arShownList.size();
 
 	*pResult = 0;
 	if (m_bNoDispUpdates || bOutOfRange)
